requirements.txt
# core
torch>=2.3
torchvision>=0.18
transformers>=4.45
accelerate>=0.33
sentencepiece>=0.2
protobuf>=4.25
Pillow>=10.3
timm>=0.9
einops>=0.7
psutil>=5.9
numpy>=1.26

# optional speed/quant
bitsandbytes>=0.43; platform_system != "Windows"
qwen-vl-utils>=0.0.8

# metrics
scikit-learn>=1.4